Initial spike
Make a robust and simple first draft of the task. In this commit I
create an exporter that scrapes (only one URL for now defined with a
const, not a flag or config) URL and exposed the status.

I develop the struct that should expose the metrics and draft the two
http functions that will get the status of the URL and the response
time. The response time is not yet ready - the function is the same as
the status function.

The final result here is to have the structure to build upon.

Things to fix and finish:
- develop the response time function
- add flag with HTTP URLs or config file/env
- make unit tests and more advanced error handling
- make more advanced logging
- add Dockerfile
- add kubernetes manifest